refactor(lab): rename constants to config and consolidate domain naming
- Renamed `src/features/lab/constants.ts` to `src/features/lab/config.ts` to better reflect mixed domain concerns (BlogDrafter and Toolbox).
- Updated imports in `BlogDrafter.tsx` and `Toolbox.tsx`.
- Verified with audit and visual regression tests.

PR: #885
