+++ src/features/lab/Toolbox.tsx
@@ -11,7 +11,7 @@ import { SearchBox } from '@/components/ui/SearchBox';
 import { EmptyState } from '@/components/ui/EmptyState';
 import { Search } from 'lucide-react';
 import { cn } from '@/lib/utils';
-import { GEAR_PILLS } from "./constants";
+import { GEAR_PILLS } from "./config";
 
 export default function Toolbox() {
   const { filteredCategories, searchTerm, setSearchTerm, view, setView, selectedPill, setSelectedPill } = useToolbox();
