+++ src/features/lab/config.ts
